Tables: Wrong SNode was assigned to GridCells
EditorComponent.findCellWithId didn't work because of that.

com.mbeddr.mpsutil.margincell stores the node and id of a cell and uses
findCellWithId to resolve that cell in an EditorComponent.

The SNode of the wrapped cell was used for the GridCell, but as they are
created by the table, it's correct are required to assign them the SNode
of the table. Cells are expected to have the SNode of the ConceptEditor
they were created in.

Fixes #128