Shadow: First "cross rewrite" references are working
A rewrite call can declare dependencies on other rewrite calls that are
then used for reference resolution.

Works like the cross model reference resolution in MPS, but shadow
models allow you to have any node as the root of a rewrite
transformation.

Currently, dependencies are automatically added for reference targets
that are of the same concept as the root of the rewrite. But this is
just to see if it works. The transformation language needs some concepts
to control the dependencies.