Shadow: Mapping labels
Mapping labels work as in MPS, but allow multiple input elements for the
key of the map entry.

They can only be used in eager mode (in a rewrite), because we need to
fill the map before we can query elements from it.

There is now a second implementation of the state machines example that
uses mapping labels.

Fixes slisson/shadowmodels#100