Shadow: Support ErrorPlaceholder for modules/models in the transformation output
When an exception occurred during the transformation of a
shadow model/module, it was replaced by an ErrorPlaceholder (as any
other node).
The next synchronisation step didn't check for that but just saw a node
without a name for which it then generated one.
Not it checks if it an ErrorPlaceholder and used the error message.