Shadow: Adapters for SModel and SModule
To support transformations on modules and models, NodeToSNodeAdapter
by default represents models/modules as nodes and there is only a single
root node for the repository.

When the incremental engine is used for other use cases, they need to be
represented as SModel/SModule. The mode can now be changed.