dropdown: fixed dropdown substitutions
The previous implementation used to set the active substitution on the last leaf, which in practice
meant that generally it was activated for the most right cell.

In this implemantion we want to use the complete editor cell from the dropdown to active the
substitution. The case we have to care for, is when in example we have no value set yet.

In that case, we have to look inside the EdiotrCell_Collection and activate the substitution for the
inner cell.