dropdown: fixed dropdown substitutions
The previous implementation used to set the active substitution on the last leaf, which in practice
meant that generally it was activated for the most right cell.

We want to active the substitution for the complete cell inside the dropdown. The only case we
have to take care is when a collection contains a single cell as in that case, we don't want to
active the substitution for the collection cell but for the inner cell.