model-api: moved ITreeChangeVisitor.nodeRemoved/.nodeAdded to new ITreeChangeVisitorEx
Allows more efficient diff after a bulk import, because new subtree
can be skipped instead of calling .nodeAdded on all new nodes.
Visitors are often not interessted in nodeAdded events, because the node
must have been added somewhere, so there is always also a
childrenChanged event and that is usually sufficient.