GrammarCells: Show popup for ambiguous INSERT/INSERT_BEFORE actions
When you press enter to insert a new node then the location where a node
is inserted is often not where you want it and it's difficult to insert
it at the correct location.

In cases where there are multiple possible locations now a popup is
shown and you can choose the location from a list.
This is consistent with the behavior of node substitiute and side
transformation actions where the code completion menu is shown if there
are multiple options.

This new behavior doesn't require any additional cells (yet). It's
enable globally for all languages.