Explicit guard for CALL with null input
Although CALL documentation states it's legitimate way to prevent template invocation, I don't think such implicit assumption is good.
Besides, it's not the case for compiled templates, unless we find a mechanism to make sure compiled and interpreted are the same, it's better not to rely on implicit assumptions.
Avoiding call with null input makes use of CALL from a Create Root template complicated (there's no input, but it's not wise to avoid use of CALL in such templates).