GrammarCells: avoid showing insertpopup with the same insert action
In some scenarios like with tables it may happen that the insert popup will be shown even if there are no ambiguous insert actions
and will suggest the same insert action multiple times (i.e. because it's registered for the grid cell and its content cell)
To avoid this, only unique actions are considered for the insert popup.