Disabled the insert action popup by default
It's more annoying than useful.
It tried to solve the problem of being able to choose where to insert a
node when the cursor it at the end of multiple collections at the same
time.
But currently the menu opens to often. A more advanced logic is
necessary automatically choose one action if multiple are applicable.

(cherry picked from commit 88bbb2df857e3419311ee08a59acfe7653c8aa6e)