celllayout.sandboxlang: add reproducer for bug
Given an indent collection containing an horizontal collection, if the
internal collection is large enought this is reported on a new line.
Now after a certain size the editor doesn't display the full node, in
fact the scroll-bar doesn't allow to scroll down to see the last part
of the node.
I suspect that the height of the resulting node it's not calculated
correctly.