fix(treenotation): incorrect position of the insert button
It was placed in the center between two subtree root, but it's more intuitive to place it at the
location where the new node will appear.
The previous solution might have been choosen to avoid overlapping buttons, but this is now
implemented properly by adding more space around children if necessary.