perf(linenumbers): they were slowing down the update of the main editor
Expensive computations were done in the paint method which are supposed to be done in the relayout
method.

The computation of the lines and their positions is now done only once after 1 second of inactivity.
Usually, the inconsistency is only noticeable for the line number at the end.
At least the line numbers are not important enough for the user experience to justify more
utilization of the EDT thread, which is already an overused resource.

Fixes #1662