feat: new "cached" cell that allows a more faine grained reuse of editor cells
For concept editors that contain cells with expensive computations, these cells can be wrapped with
the "cached" cell. It tracks the dependencies independently for this cell and reuses it when the
concept editor is invoked again.

When a cell is updated all its ancestors are also updated. The closer a node is to the root of the
cell tree, the more often it is re-computed and the more critical it is not to put expensive
computations into it.