feat: asynchronous editor updates
Usually, you have to be very careful when traversing the model inside editor code because it
introduces dependencies between cells and model elements. A small change in the model can result
in large updates in the editor. For large root nodes the editor can become almost unusable.

The proposed solution is to delay updates of editor cells that are currently not important for the
edited part of the root node and tolerate a short inconsistency.
Most of the time the dependency is either unecessary because MPS just doesn't track them
fine-grained enough, or the change has no effect on the final result, or the affected cells aren't
currently visible because they are outside the viewport.

The cells that are updated immediately are currently determined by which cell is selected.
This heuristic might need to be tweaked after collecting some experience.

There is a new action `Asynchronous Editor Updates` in the context menu of the editor.